Сделал работу с файлами у вакансий

+++ src/main/java/ru/job4j/dreamjob/service/SimpleVacancyService.java
@@ -1,6 +1,8 @@
 package ru.job4j.dreamjob.service;
 
 import org.springframework.stereotype.Service;
+import ru.job4j.dreamjob.dto.FileDto;
+import ru.job4j.dreamjob.model.File;
 import ru.job4j.dreamjob.model.Vacancy;
 import ru.job4j.dreamjob.repository.VacancyRepository;
 
@@ -10,24 +12,46 @@ import java.util.Optional;
 @Service
 public class SimpleVacancyService implements VacancyService {
     private final VacancyRepository vacancyRepository;
+    private final FileService fileService;
 
-    public SimpleVacancyService(VacancyRepository vacancyRepository) {
+    public SimpleVacancyService(VacancyRepository vacancyRepository, FileService fileService) {
         this.vacancyRepository = vacancyRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Vacancy save(Vacancy vacancy) {
+    public Vacancy save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
         return vacancyRepository.save(vacancy);
+    }
+
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        File file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-        return vacancyRepository.deleteById(id);
+        Optional<Vacancy> vacancyOptional = findById(id);
+        if (vacancyOptional.isEmpty()) {
+            return false;
+        }
+        boolean isDeleted = vacancyRepository.deleteById(id);
+        fileService.deleteById(vacancyOptional.get().getFileId());
+        return isDeleted;
     }
 
     @Override
-    public boolean update(Vacancy vacancy) {
-        return vacancyRepository.update(vacancy);
+    public boolean update(Vacancy vacancy, FileDto image) {
+        boolean isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty){
+            return vacancyRepository.update(vacancy);
+        }
+        int oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        boolean isUpdated = vacancyRepository.update(vacancy);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
